==> sensitivity-analysis/scripts/4c-dementia-pills.ipynb <==
# Dementia and Pills Plots

import pandas as pd
import numpy as np

from matplotlib_venn import venn2, venn2_circles, venn2_unweighted
from matplotlib_venn import venn3, venn3_circles
from matplotlib import pyplot as plt
%matplotlib inline
plt.rcParams['figure.figsize'] = [15, 7]

import stata_setup

stata_setup.config("/Applications/Stata", "mp")

%%stata
clear

%%stata
use "../data/step-1-self/cleaned-data.dta"

svyset w6varunit [pweight=w6anfinwgt0], strata(w6varstrat)

recode is6prxyrelat (6/30 32/92 =6), gen (prxyrelat_new)
label variable prxyrelat_new "proxy-relationship"
label define prxyrelat 2 "spouse/partner" 3 "daughter" 4 "son" 5 "daughter-in-law" 31 "paid aide/employee" 6 "other"
label values prxyrelat_new prxyrelat

label variable is6famrrutin "how familiar with routine"
label define is6famrrutin 1 "very familiar" 2 "somewhat familiar" 3 "a little familiar" 4 "not at all/refused/don't know"
label values is6famrrutin is6famrrutin

%%stata

gen dementia_diagnosis=.
replace dementia_diagnosis=0 if hc6disescn9==0
replace dementia_diagnosis=1 if hc6disescn9==7
replace dementia_diagnosis=1 if hc6disescn9==1

label variable dementia_diagnosis "diagnosed with dementia"
label define dementia_diagnosis_lab 0 "not diagnosed" 1 "diagnosed"
label values dementia_diagnosis dementia_diagnosis_lab

%%stata
gen pillsmax_clean = .
replace pillsmax_clean=ma6pillsmax if ma6pillsmax>0

quietly: tabout pillsmax_clean if subpop using "../tables/plottables/pillsmax.csv", svy cells(cells) format(2) replace
    
quietly: tabout dementia_diagnosis pillsmax_clean if subpop using "../tables/plottables/diagnosis-pills.csv", svy cells(row) format(2) replace
quietly: tabout dementia_diagnosis pillsmax_clean if subpop using "../tables/plottables/diagnosis-pills-2.csv", svy cells(cells) format(2) replace

pills = pd.read_table('../tables/plottables/pillsmax.csv')
pills

diagnosis_pills_row = pd.read_table('../tables/plottables/diagnosis-pills.csv', skiprows=1)
diagnosis_pills_row = diagnosis_pills_row.drop([0])
diagnosis_pills_row

diagnosis_pills_cell = pd.read_table('../tables/plottables/diagnosis-pills-2.csv', skiprows=1)
diagnosis_pills_cell = diagnosis_pills_cell.drop([0])
diagnosis_pills_cell

%%stata
quietly: tabout dementia pillsmax_clean if subpop using "../tables/plottables/dementia-pills.csv", svy cells(row) format(2) replace
quietly: tabout dementia pillsmax_clean if subpop using "../tables/plottables/dementia-pills-2.csv", svy cells(cells) format(2) replace

dementia_pills_row = pd.read_table('../tables/plottables/dementia-pills.csv', skiprows=1)
dementia_pills_row = dementia_pills_row.drop([0])
dementia_pills_row

dementia_pills_cell = pd.read_table('../tables/plottables/dementia-pills-2.csv', skiprows=1)
dementia_pills_cell = dementia_pills_cell.drop([0])
dementia_pills_cell

pills.to_csv('../tables/clean-csv/pills.csv')

diagnosis_pills_row.to_csv('../tables/clean-csv/diagnosis-pills-row.csv')
diagnosis_pills_cell.to_csv('../tables/clean-csv/diagnosis-pills-cell.csv')
dementia_pills_row.to_csv('../tables/clean-csv/dementia-pills-row.csv')
dementia_pills_cell.to_csv('../tables/clean-csv/dementia-pills-cell.csv')
